+++ src/SelectionManager.ts
@@ -6,7 +6,7 @@
 import { ITerminal, ISelectionManager, IBuffer, CharData, IBufferLine } from './Types';
 import { XtermListener } from './common/Types';
 import { MouseHelper } from './utils/MouseHelper';
-import * as Browser from './shared/utils/Browser';
+import * as Browser from './core/Platform';
 import { CharMeasure } from './ui/CharMeasure';
 import { EventEmitter } from './common/EventEmitter';
 import { SelectionModel } from './SelectionModel';
