+++ src/core/Platform.ts
