+++ src/renderer/ColorManager.ts
@@ -3,8 +3,7 @@
  * @license MIT
  */
 
-import { IColorManager } from './Types';
-import { IColor, IColorSet } from '../shared/Types';
+import { IColorManager, IColor, IColorSet } from './Types';
 import { ITheme } from 'xterm';
 
 const DEFAULT_FOREGROUND = fromHex('#ffffff');
